infrastructure/pyproject.toml
[tool.poetry]
name = "infrastructure"
version = "0.1.0"
description = "Infrastructure for Raktar website"
authors = ["David Steiner"]

[tool.poetry.dependencies]
python = "~3.10"
aws-cdk-lib = "^2.84.0"
cdk-static-site = "^0.1.0"

[tool.poetry.dev-dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
